refactor: Move routes to single file in endpoints folder
