--- src/routes/createSession.ts
@@ -1,9 +1,0 @@
-import { Route } from '../definitions';
-import { createSession } from '../handlers';
-
-const createSessionRoute: Route = {
-  handler: createSession,
-  url: '/createSession',
-};
-
-export default createSessionRoute;
