--- src/routes/joinSession.ts
@@ -1,9 +1,0 @@
-import { Route } from '../definitions';
-import { joinSession } from '../handlers';
-
-const joinSessionRoute: Route = {
-  handler: joinSession,
-  url: '/joinSession',
-};
-
-export default joinSessionRoute;
